backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-filter
django-cors-headers
drf-spectacular
celery[redis]>=5.3
django-celery-beat
django-celery-results
psycopg[binary]>=3.1
redis>=5.0
python-decouple
gunicorn

# Scraping
pandas
SwimScraper
swimrankingsscraper
scrapy
selenium
webdriver-manager
openpyxl
playwright
playwright-stealth
